Date and time support in rules' expressions (MathJS doesn't support dates, so it was extended with using date/time types from the js-joda library) setup.sh: fixed running on MacOS update.sh: added prompt before doing git hard reset BE: config-service: fixed fetching application list (old non empty status displayed as empty configurations)

+++ scripts/update.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
-# 
+#
 # Copyright 2021 Google LLC
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
@@ -13,8 +13,20 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+
+RED='\033[0;31m' # Red Color
+NC='\033[0m' # No Color
+
+echo -e "${RED}WARNING: all local changes will be overwritten${NC}"
+read -p "Do you want to update repository and deployed app? (Y/N): " -n 1 -r
+echo    # (optional) move to a new line
+if [[ $REPLY =~ ^[Nn]$ ]]
+then
+    exit
+fi
+
 cd ..
-git fetch 
+git fetch
 git reset --hard origin/main
 
 PROJECT_ID=$(gcloud config get-value project 2> /dev/null)
